@@ -5,11 +5,11 @@
     "keywords": ["xbbcode","bbcode"],
     "license": "GPL-2.0-or-later",
     "require": {
-        "php": ">=7.4",
+        "php": ">=8.0.2",
         "geshi/geshi": "^1.0.9.1"
     },
     "require-dev": {
-        "phpunit/phpunit": "^9.5",
+        "phpunit/phpunit": "^9.6",
         "friendsofphp/php-cs-fixer": "^3.0"
     },
     "authors": [
